rename rest of TabItemList to TabList
