+++ packages/react/src/components/Tabs/Tab/Tab.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 
-import { TabList } from '../TabItemList';
+import { TabList } from '../TabList';
 import { Tabs } from '..';
 
 import { Tab } from '.';
