+++ packages/react/src/components/Tabs/TabList/TabList.module.css
